+++ p6/p6.cc
@@ -2,8 +2,8 @@
 using namespace std;
 
 int main() {
-  long long sumSquares = 0;
-  long long squareSums = 0;
+  long sumSquares = 0;
+  long squareSums = 0;
   for(int i = 1; i <=100; i++) {
     squareSums+=i;
     sumSquares += i*i;
